Refactor Java
