+++ java/test/integration/cluster/BUILD
@@ -18,33 +18,8 @@
 load("@typedb_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
 load("@typedb_dependencies//builder/java:rules.bzl", "typedb_java_test")
 
-exports_files(
-    ["ExampleTest.java"],
-    visibility = ["//java:__subpackages__"],
-)
-
-typedb_java_test(
-    name = "test-example",
-    srcs = ["ExampleTest.java"],
-    server_artifacts = {
-        "@typedb_bazel_distribution//platform:is_linux_arm64": "@typedb_artifact_linux-arm64//file",
-        "@typedb_bazel_distribution//platform:is_linux_x86_64": "@typedb_artifact_linux-x86_64//file",
-        "@typedb_bazel_distribution//platform:is_mac_arm64": "@typedb_artifact_mac-arm64//file",
-        "@typedb_bazel_distribution//platform:is_mac_x86_64": "@typedb_artifact_mac-x86_64//file",
-#        "@typedb_bazel_distribution//platform:is_windows_x86_64": "@typedb_artifact_windows-x86_64//file",
-    },
-    test_class = "com.typedb.driver.test.integration.cloud.ExampleTest",
-    deps = [
-        # Internal dependencies
-        "//java:driver-java",
-        "//java/api",
-        "//java/common",
-
-        # External dependencies from @typedb
-        "@maven//:org_slf4j_slf4j_api",
-#        "@maven//:com_typedb_typedb_runner",
-    ],
-)
+# TODO: This package can contain cluster-specific integration tests. Tests available for running for both community
+# edition and cluster should be run for both setups (and be limited: always prefer BDDs to integration tests)!
 
 checkstyle_test(
     name = "checkstyle",
