+++ python/tests/integration/cluster/BUILD
@@ -18,12 +18,9 @@
 load(":cluster_test_rule.bzl", "typedb_cloud_py_test")
 load("@typedb_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
 load("@typedb_driver_pip//:requirements.bzl", "requirement")
-load("@rules_python//python:defs.bzl", "py_test")
 
-exports_files(
-    ["test_example.py"],
-    visibility = ["//python:__subpackages__"],
-)
+# TODO: This package can contain cluster-specific integration tests. Tests available for running for both community
+# edition and cluster should be run for both setups (and be limited: always prefer BDDs to integration tests)!
 
 #typedb_cloud_py_test(
 #    name = "test_failover",
